+++ e2e/runner/runner.go
@@ -62,15 +62,17 @@ type E2ERunner struct {
 	EVMAuth  *bind.TransactOpts
 	ZEVMAuth *bind.TransactOpts
 
-	// contracts
-	ZetaEthAddr          ethcommon.Address
-	ZetaEth              *zetaeth.ZetaEth
-	ConnectorEthAddr     ethcommon.Address
-	ConnectorEth         *zetaconnectoreth.ZetaConnectorEth
-	ERC20CustodyAddr     ethcommon.Address
-	ERC20Custody         *erc20custody.ERC20Custody
-	ERC20Addr            ethcommon.Address
-	ERC20                *erc20.ERC20
+	// contracts evm
+	ZetaEthAddr      ethcommon.Address
+	ZetaEth          *zetaeth.ZetaEth
+	ConnectorEthAddr ethcommon.Address
+	ConnectorEth     *zetaconnectoreth.ZetaConnectorEth
+	ERC20CustodyAddr ethcommon.Address
+	ERC20Custody     *erc20custody.ERC20Custody
+	ERC20Addr        ethcommon.Address
+	ERC20            *erc20.ERC20
+
+	// contracts zevm
 	ERC20ZRC20Addr       ethcommon.Address
 	ERC20ZRC20           *zrc20.ZRC20
 	ETHZRC20Addr         ethcommon.Address
@@ -85,14 +87,13 @@ type E2ERunner struct {
 	ConnectorZEVM        *connectorzevm.ZetaConnectorZEVM
 	WZetaAddr            ethcommon.Address
 	WZeta                *wzeta.WETH9
-
-	TestDAppAddr       ethcommon.Address
-	ZEVMSwapAppAddr    ethcommon.Address
-	ZEVMSwapApp        *zevmswap.ZEVMSwapApp
-	ContextAppAddr     ethcommon.Address
-	ContextApp         *contextapp.ContextApp
-	SystemContractAddr ethcommon.Address
-	SystemContract     *systemcontract.SystemContract
+	TestDAppAddr         ethcommon.Address
+	ZEVMSwapAppAddr      ethcommon.Address
+	ZEVMSwapApp          *zevmswap.ZEVMSwapApp
+	ContextAppAddr       ethcommon.Address
+	ContextApp           *contextapp.ContextApp
+	SystemContractAddr   ethcommon.Address
+	SystemContract       *systemcontract.SystemContract
 
 	// config
 	CctxTimeout    time.Duration