+++ x/crosschain/keeper/evm_deposit_test.go
@@ -117,6 +117,152 @@ func TestMsgServer_HandleEVMDeposit(t *testing.T) {
 		fungibleMock.AssertExpectations(t)
 	})
 
+	t.Run("should error on processing ERC20 deposit calling fungible method for contract call if process logs fails", func(t *testing.T) {
+		k, ctx, _, _ := keepertest.CrosschainKeeperWithMocks(t, keepertest.CrosschainMockOptions{
+			UseFungibleMock: true,
+		})
+
+		senderChain := getValidEthChainID(t)
+
+		fungibleMock := keepertest.GetCrosschainFungibleMock(t, k)
+		receiver := sample.EthAddress()
+		amount := big.NewInt(42)
+
+		// expect DepositCoinZeta to be called
+		// ZRC20DepositAndCallContract(ctx, from, to, msg.Amount.BigInt(), senderChain, msg.Message, contract, data, msg.CoinType, msg.Asset)
+		fungibleMock.On(
+			"ZRC20DepositAndCallContract",
+			ctx,
+			mock.Anything,
+			receiver,
+			amount,
+			senderChain,
+			mock.Anything,
+			coin.CoinType_ERC20,
+			mock.Anything,
+		).Return(&evmtypes.MsgEthereumTxResponse{
+			Logs: []*evmtypes.Log{
+				{
+					Address:     receiver.Hex(),
+					Topics:      []string{},
+					Data:        []byte{},
+					BlockNumber: uint64(ctx.BlockHeight()),
+					TxHash:      sample.Hash().Hex(),
+					TxIndex:     1,
+					BlockHash:   sample.Hash().Hex(),
+					Index:       1,
+				},
+			},
+		}, true, nil)
+
+		fungibleMock.On("GetSystemContract", mock.Anything).Return(fungibletypes.SystemContract{}, false)
+
+		// call HandleEVMDeposit
+		cctx := sample.CrossChainTx(t, "foo")
+		cctx.InboundTxParams.TxOrigin = ""
+		cctx.GetCurrentOutTxParam().Receiver = receiver.String()
+		cctx.GetInboundTxParams().Amount = math.NewUintFromBigInt(amount)
+		cctx.GetInboundTxParams().CoinType = coin.CoinType_ERC20
+		cctx.GetInboundTxParams().Sender = sample.EthAddress().String()
+		cctx.GetInboundTxParams().SenderChainId = senderChain
+		cctx.RelayedMessage = ""
+		cctx.GetInboundTxParams().Asset = ""
+		reverted, err := k.HandleEVMDeposit(
+			ctx,
+			cctx,
+		)
+		require.Error(t, err)
+		require.False(t, reverted)
+		fungibleMock.AssertExpectations(t)
+	})
+
+	t.Run("can process ERC20 deposit calling fungible method for contract call if process logs doesnt fail", func(t *testing.T) {
+		k, ctx, _, _ := keepertest.CrosschainKeeperWithMocks(t, keepertest.CrosschainMockOptions{
+			UseFungibleMock: true,
+		})
+
+		senderChain := getValidEthChainID(t)
+
+		fungibleMock := keepertest.GetCrosschainFungibleMock(t, k)
+		receiver := sample.EthAddress()
+		amount := big.NewInt(42)
+
+		// expect DepositCoinZeta to be called
+		// ZRC20DepositAndCallContract(ctx, from, to, msg.Amount.BigInt(), senderChain, msg.Message, contract, data, msg.CoinType, msg.Asset)
+		fungibleMock.On(
+			"ZRC20DepositAndCallContract",
+			ctx,
+			mock.Anything,
+			receiver,
+			amount,
+			senderChain,
+			mock.Anything,
+			coin.CoinType_ERC20,
+			mock.Anything,
+		).Return(&evmtypes.MsgEthereumTxResponse{
+			Logs: []*evmtypes.Log{
+				{
+					Address:     receiver.Hex(),
+					Topics:      []string{},
+					Data:        []byte{},
+					BlockNumber: uint64(ctx.BlockHeight()),
+					TxHash:      sample.Hash().Hex(),
+					TxIndex:     1,
+					BlockHash:   sample.Hash().Hex(),
+					Index:       1,
+				},
+			},
+		}, true, nil)
+
+		fungibleMock.On("GetSystemContract", mock.Anything).Return(fungibletypes.SystemContract{
+			ConnectorZevm: sample.EthAddress().Hex(),
+		}, true)
+
+		// call HandleEVMDeposit
+		cctx := sample.CrossChainTx(t, "foo")
+		cctx.InboundTxParams.TxOrigin = ""
+		cctx.GetCurrentOutTxParam().Receiver = receiver.String()
+		cctx.GetInboundTxParams().Amount = math.NewUintFromBigInt(amount)
+		cctx.GetInboundTxParams().CoinType = coin.CoinType_ERC20
+		cctx.GetInboundTxParams().Sender = sample.EthAddress().String()
+		cctx.GetInboundTxParams().SenderChainId = senderChain
+		cctx.RelayedMessage = ""
+		cctx.GetInboundTxParams().Asset = ""
+		reverted, err := k.HandleEVMDeposit(
+			ctx,
+			cctx,
+		)
+		require.NoError(t, err)
+		require.False(t, reverted)
+		fungibleMock.AssertExpectations(t)
+	})
+
+	t.Run("should error if invalid sender", func(t *testing.T) {
+		k, ctx, _, _ := keepertest.CrosschainKeeperWithMocks(t, keepertest.CrosschainMockOptions{
+			UseFungibleMock: true,
+		})
+
+		receiver := sample.EthAddress()
+		amount := big.NewInt(42)
+
+		// call HandleEVMDeposit
+		cctx := sample.CrossChainTx(t, "foo")
+		cctx.InboundTxParams.TxOrigin = ""
+		cctx.GetCurrentOutTxParam().Receiver = receiver.String()
+		cctx.GetInboundTxParams().Amount = math.NewUintFromBigInt(amount)
+		cctx.GetInboundTxParams().CoinType = coin.CoinType_ERC20
+		cctx.GetInboundTxParams().Sender = "invalid"
+		cctx.GetInboundTxParams().SenderChainId = 987
+		cctx.RelayedMessage = ""
+		cctx.GetInboundTxParams().Asset = ""
+		reverted, err := k.HandleEVMDeposit(
+			ctx,
+			cctx,
+		)
+		require.Error(t, err)
+		require.False(t, reverted)
+	})
+
 	t.Run("should return error with non-reverted if deposit ERC20 fails with tx non-failed", func(t *testing.T) {
 		k, ctx, _, _ := keepertest.CrosschainKeeperWithMocks(t, keepertest.CrosschainMockOptions{
 			UseFungibleMock: true,
@@ -366,6 +512,10 @@ func TestMsgServer_HandleEVMDeposit(t *testing.T) {
 
 		data, err := hex.DecodeString("DEADBEEF")
 		require.NoError(t, err)
+		cctx := sample.CrossChainTx(t, "foo")
+		b, err := cctx.Marshal()
+		require.NoError(t, err)
+		ctx = ctx.WithTxBytes(b)
 		fungibleMock.On(
 			"ZRC20DepositAndCallContract",
 			ctx,
@@ -378,7 +528,6 @@ func TestMsgServer_HandleEVMDeposit(t *testing.T) {
 			mock.Anything,
 		).Return(&evmtypes.MsgEthereumTxResponse{}, false, nil)
 
-		cctx := sample.CrossChainTx(t, "foo")
 		cctx.GetCurrentOutTxParam().Receiver = sample.EthAddress().String()
 		cctx.GetInboundTxParams().Amount = math.NewUintFromBigInt(amount)
 		cctx.GetInboundTxParams().CoinType = coin.CoinType_ERC20
@@ -393,6 +542,7 @@ func TestMsgServer_HandleEVMDeposit(t *testing.T) {
 		require.NoError(t, err)
 		require.False(t, reverted)
 		fungibleMock.AssertExpectations(t)
+		require.Equal(t, uint64(ctx.BlockHeight()), cctx.GetCurrentOutTxParam().OutboundTxObservedExternalHeight)
 	})
 
 	t.Run("should deposit into receiver with specified data if no address parsed with data", func(t *testing.T) {