+++ x/crosschain/keeper/grpc_query_last_block_height_test.go
@@ -1,20 +1,22 @@
-package keeper
+package keeper_test
 
 import (
+	"math"
 	"testing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
 	"github.com/stretchr/testify/require"
+	keepertest "github.com/zeta-chain/zetacore/testutil/keeper"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
 func TestLastBlockHeightQuerySingle(t *testing.T) {
-	keeper, ctx := setupKeeper(t)
+	k, ctx, _, _ := keepertest.CrosschainKeeper(t)
 	wctx := sdk.WrapSDKContext(ctx)
-	msgs := createNLastBlockHeight(keeper, ctx, 2)
+	msgs := createNLastBlockHeight(k, ctx, 2)
 	for _, tc := range []struct {
 		desc     string
 		request  *types.QueryGetLastBlockHeightRequest
@@ -43,7 +45,7 @@ func TestLastBlockHeightQuerySingle(t *testing.T) {
 	} {
 		tc := tc
 		t.Run(tc.desc, func(t *testing.T) {
-			response, err := keeper.LastBlockHeight(wctx, tc.request)
+			response, err := k.LastBlockHeight(wctx, tc.request)
 			if tc.err != nil {
 				require.ErrorIs(t, err, tc.err)
 			} else {
@@ -53,10 +55,43 @@ func TestLastBlockHeightQuerySingle(t *testing.T) {
 	}
 }
 
+func TestLastBlockHeightLimits(t *testing.T) {
+	t.Run("should err if last send height is max int", func(t *testing.T) {
+		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
+		wctx := sdk.WrapSDKContext(ctx)
+		k.SetLastBlockHeight(ctx, types.LastBlockHeight{
+			Index:          "index",
+			LastSendHeight: math.MaxInt64,
+		})
+
+		res, err := k.LastBlockHeight(wctx, &types.QueryGetLastBlockHeightRequest{
+			Index: "index",
+		})
+		require.Nil(t, res)
+		require.Error(t, err)
+	})
+
+	t.Run("should err if last receive height is max int", func(t *testing.T) {
+		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
+		wctx := sdk.WrapSDKContext(ctx)
+		k.SetLastBlockHeight(ctx, types.LastBlockHeight{
+			Index:             "index",
+			LastSendHeight:    10,
+			LastReceiveHeight: math.MaxInt64,
+		})
+
+		res, err := k.LastBlockHeight(wctx, &types.QueryGetLastBlockHeightRequest{
+			Index: "index",
+		})
+		require.Nil(t, res)
+		require.Error(t, err)
+	})
+}
+
 func TestLastBlockHeightQueryPaginated(t *testing.T) {
-	keeper, ctx := setupKeeper(t)
+	k, ctx, _, _ := keepertest.CrosschainKeeper(t)
 	wctx := sdk.WrapSDKContext(ctx)
-	msgs := createNLastBlockHeight(keeper, ctx, 5)
+	msgs := createNLastBlockHeight(k, ctx, 5)
 
 	request := func(next []byte, offset, limit uint64, total bool) *types.QueryAllLastBlockHeightRequest {
 		return &types.QueryAllLastBlockHeightRequest{
@@ -71,7 +106,7 @@ func TestLastBlockHeightQueryPaginated(t *testing.T) {
 	t.Run("ByOffset", func(t *testing.T) {
 		step := 2
 		for i := 0; i < len(msgs); i += step {
-			resp, err := keeper.LastBlockHeightAll(wctx, request(nil, uint64(i), uint64(step), false))
+			resp, err := k.LastBlockHeightAll(wctx, request(nil, uint64(i), uint64(step), false))
 			require.NoError(t, err)
 			for j := i; j < len(msgs) && j < i+step; j++ {
 				require.Equal(t, &msgs[j], resp.LastBlockHeight[j-i])
@@ -82,7 +117,7 @@ func TestLastBlockHeightQueryPaginated(t *testing.T) {
 		step := 2
 		var next []byte
 		for i := 0; i < len(msgs); i += step {
-			resp, err := keeper.LastBlockHeightAll(wctx, request(next, 0, uint64(step), false))
+			resp, err := k.LastBlockHeightAll(wctx, request(next, 0, uint64(step), false))
 			require.NoError(t, err)
 			for j := i; j < len(msgs) && j < i+step; j++ {
 				require.Equal(t, &msgs[j], resp.LastBlockHeight[j-i])
@@ -91,12 +126,12 @@ func TestLastBlockHeightQueryPaginated(t *testing.T) {
 		}
 	})
 	t.Run("Total", func(t *testing.T) {
-		resp, err := keeper.LastBlockHeightAll(wctx, request(nil, 0, 0, true))
+		resp, err := k.LastBlockHeightAll(wctx, request(nil, 0, 0, true))
 		require.NoError(t, err)
 		require.Equal(t, len(msgs), int(resp.Pagination.Total))
 	})
 	t.Run("InvalidRequest", func(t *testing.T) {
-		_, err := keeper.LastBlockHeightAll(wctx, nil)
+		_, err := k.LastBlockHeightAll(wctx, nil)
 		require.ErrorIs(t, err, status.Error(codes.InvalidArgument, "invalid request"))
 	})
 }