+++ x/observer/keeper/msg_server_reset_chain_nonces.go
@@ -10,7 +10,6 @@ import (
 )
 
 // ResetChainNonces handles resetting chain nonces
-// Authorized: admin policy group 2 (admin update)
 func (k msgServer) ResetChainNonces(goCtx context.Context, msg *types.MsgResetChainNonces) (*types.MsgResetChainNoncesResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 	if !k.GetAuthorityKeeper().IsAuthorized(ctx, msg.Creator, authoritytypes.PolicyType_groupOperational) {