+++ x/observer/keeper/tss_funds_migrator_test.go
@@ -12,10 +12,16 @@ func TestKeeper_GetTssFundMigrator(t *testing.T) {
 	t.Run("Successfully set funds migrator for chain", func(t *testing.T) {
 		k, ctx, _, _ := keepertest.ObserverKeeper(t)
 		chain := sample.TssFundsMigrator(1)
+		_, found := k.GetFundMigrator(ctx, chain.ChainId)
+		require.False(t, found)
 		k.SetFundMigrator(ctx, chain)
 		tfm, found := k.GetFundMigrator(ctx, chain.ChainId)
 		require.True(t, found)
 		require.Equal(t, chain, tfm)
+
+		k.RemoveAllExistingMigrators(ctx)
+		_, found = k.GetFundMigrator(ctx, chain.ChainId)
+		require.False(t, found)
 	})
 	t.Run("Verify only one migrator can be created for a chain", func(t *testing.T) {
 		k, ctx, _, _ := keepertest.ObserverKeeper(t)
@@ -28,5 +34,4 @@ func TestKeeper_GetTssFundMigrator(t *testing.T) {
 		require.Equal(t, 1, len(migratorList))
 		require.Equal(t, tfm2, migratorList[0])
 	})
-
 }